+++ src/test/java/rules/BaseballRulesTest.java
@@ -1,4 +1,4 @@
-package utils;
+package rules;
 
 import static org.assertj.core.api.Assertions.assertThat;
